+++ gridcapa-task-manager-app/src/test/java/com/farao_community/farao/gridcapa/task_manager/app/TaskManagerTest.java
@@ -11,7 +11,6 @@ import com.farao_community.farao.gridcapa.task_manager.app.entities.ProcessEvent
 import com.farao_community.farao.gridcapa.task_manager.app.entities.Task;
 import com.farao_community.farao.minio_adapter.starter.MinioAdapter;
 import com.farao_community.farao.minio_adapter.starter.MinioAdapterConstants;
-import io.minio.messages.Event;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -22,13 +21,10 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.cloud.stream.function.StreamBridge;
 
 import java.time.OffsetDateTime;
-import java.time.ZoneOffset;
-import java.util.Iterator;
 import java.util.UUID;
 
-import static com.farao_community.farao.gridcapa.task_manager.api.TaskStatus.*;
+import static com.farao_community.farao.gridcapa.task_manager.api.TaskStatus.RUNNING;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
@@ -43,6 +39,7 @@ class TaskManagerTest {
 
     @MockBean
     private StreamBridge streamBridge; // Useful to avoid AMQP connection that would fail
+
     @MockBean
     private MinioAdapter minioAdapter;
 
@@ -59,179 +56,6 @@ class TaskManagerTest {
     void cleanDatabase() {
         taskRepository.deleteAll();
         processFileRepository.deleteAll();
-    }
-
-    @Test
-    void testUpdate() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-09-30T21:00Z");
-        Event event = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        taskManager.updateTasks(event);
-
-        assertTrue(taskRepository.findByTimestamp(taskTimestamp).isPresent());
-        Task task = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals("CSE/D2CC/CGMs/cgm-test", task.getInput("CGM").orElseThrow().getFileObjectKey());
-        assertEquals("cgm-test", task.getInput("CGM").orElseThrow().getFilename());
-    }
-
-    @Test
-    void testUpdateWithTwoFileTypesInTheSameTimestamp() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-09-30T21:00Z");
-        Event eventCgm = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        Event eventCrac = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CRAC", "CSE/D2CC/CRACs/crac-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        taskManager.updateTasks(eventCgm);
-        taskManager.updateTasks(eventCrac);
-
-        assertEquals(1, taskRepository.findAll().size());
-        assertTrue(taskRepository.findByTimestamp(taskTimestamp).isPresent());
-        Task task = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals("CSE/D2CC/CGMs/cgm-test", task.getInput("CGM").orElseThrow().getFileObjectKey());
-        assertEquals("CSE/D2CC/CRACs/crac-test", task.getInput("CRAC").orElseThrow().getFileObjectKey());
-        assertEquals("cgm-test", task.getInput("CGM").orElseThrow().getFilename());
-        assertEquals("crac-test", task.getInput("CRAC").orElseThrow().getFilename());
-    }
-
-    @Test
-    void testUpdateWithEmptyIntervalFileType() {
-        testTimeInterval("CSE_D2CC", "", 0);
-    }
-
-    @Test
-    void testUpdateWithDailyFile() {
-        testTimeInterval("CSE_D2CC", "2021-09-30T22:00Z/2021-10-01T22:00Z", 24);
-    }
-
-    private void testTimeInterval(String process, String interval, int expectedFileNumber) {
-        Event event = TaskManagerTestUtil.createEvent(minioAdapter, process, INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", interval);
-
-        taskManager.updateTasks(event);
-
-        assertEquals(expectedFileNumber, taskRepository.findAll().size());
-    }
-
-    @Test
-    void checkStatusUpdateToReady() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-09-30T21:00Z");
-        Event eventCgm = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        Event eventCrac = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CRAC", "CSE/D2CC/CRACs/crac-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        taskManager.updateTasks(eventCgm);
-        assertEquals(CREATED, taskRepository.findByTimestamp(taskTimestamp).orElseThrow().getStatus());
-        taskManager.updateTasks(eventCrac);
-        assertEquals(READY, taskRepository.findByTimestamp(taskTimestamp).orElseThrow().getStatus());
-    }
-
-    @Test
-    void checkStatusUpdateBackToCreated() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-09-30T21:00Z");
-        Event eventCgm = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-        Event eventCrac = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CRAC", "CSE/D2CC/CRACs/crac-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        taskManager.updateTasks(eventCgm);
-        taskManager.updateTasks(eventCrac);
-        taskManager.removeProcessFile(eventCrac);
-        assertEquals(CREATED, taskRepository.findByTimestamp(taskTimestamp).orElseThrow().getStatus());
-    }
-
-    @Test
-    void testCreationEventsForTwoFilesWithDifferentTypesAndSameTs() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-09-30T21:00Z");
-        Event eventCgm = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-        Event eventCrac = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CRAC", "CSE/D2CC/CRACs/crac-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        taskManager.updateTasks(eventCgm);
-        taskManager.updateTasks(eventCrac);
-
-        Task task = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals(2, task.getProcessEvents().size());
-        Iterator<ProcessEvent> processEventIterator = task.getProcessEvents().iterator();
-        ProcessEvent event = processEventIterator.next();
-        assertEquals("INFO", event.getLevel());
-        assertEquals("The CRAC : 'crac-test' is available", event.getMessage());
-        event = processEventIterator.next();
-        assertEquals("INFO", event.getLevel());
-        assertEquals("The CGM : 'cgm-test' is available", event.getMessage());
-    }
-
-    @Test
-    void testUpdateEventsForTwoFilesWithSameTypeAndSameTs() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-09-30T21:00Z");
-        Event eventCgm = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-        Event eventCgmNew = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CGM", "CSE/D2CC/CGMs/cgm-new-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-
-        taskManager.updateTasks(eventCgm);
-        taskManager.updateTasks(eventCgmNew);
-
-        Task task = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals(2, task.getProcessEvents().size());
-        Iterator<ProcessEvent> processEventIterator = task.getProcessEvents().iterator();
-        ProcessEvent event1 = processEventIterator.next();
-        ProcessEvent event2 = processEventIterator.next();
-        assertTrue(event2.getTimestamp().isBefore(event1.getTimestamp()));
-        assertEquals("INFO", event1.getLevel());
-        assertEquals("INFO", event2.getLevel());
-        assertEquals("The CGM : 'cgm-test' is available", event2.getMessage());
-        assertEquals("A new version of CGM is available : 'cgm-new-test'", event1.getMessage());
-    }
-
-    @Test
-    void testDeletionEventsOfTaskWithTwoDifferentFileTypes() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-10-01T21:00Z").withOffsetSameInstant(ZoneOffset.UTC);
-        Task task = new Task(taskTimestamp);
-
-        task.addProcessEvent(OffsetDateTime.now(), "INFO", "CGM available");
-        task.addProcessFile(
-                "CSE/D2CC/CGMs/cgm-test",
-                MinioAdapterConstants.DEFAULT_GRIDCAPA_INPUT_GROUP_METADATA_VALUE,
-                "CGM",
-                OffsetDateTime.parse("2021-10-01T21:00Z"),
-                OffsetDateTime.parse("2021-10-01T22:00Z"),
-                OffsetDateTime.now());
-
-        task.addProcessEvent(OffsetDateTime.now(), "INFO", "Crac available");
-        task.addProcessFile(
-                "CSE/D2CC/CRACs/crac-test",
-                MinioAdapterConstants.DEFAULT_GRIDCAPA_INPUT_GROUP_METADATA_VALUE,
-                "CRAC",
-                OffsetDateTime.parse("2021-10-01T21:00Z"),
-                OffsetDateTime.parse("2021-10-01T22:00Z"),
-                OffsetDateTime.now());
-
-        taskRepository.save(task);
-
-        Event eventCracDeletion = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CRAC", "CSE/D2CC/CRACs/crac-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-        taskManager.removeProcessFile(eventCracDeletion);
-
-        Task updatedTask = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals(3, updatedTask.getProcessEvents().size());
-
-        Iterator<ProcessEvent> processEventIterator = updatedTask.getProcessEvents().iterator();
-        ProcessEvent event = processEventIterator.next();
-        assertEquals("The CRAC : 'crac-test' is deleted", event.getMessage());
-        assertEquals(1, updatedTask.getProcessFiles().size());
-    }
-
-    @Test
-    void testDeletionEventsWithTaskDeletion() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-10-01T21:00Z");
-        Task task = new Task(taskTimestamp);
-        task.addProcessEvent(OffsetDateTime.now(), "INFO", "Crac available");
-        task.addProcessFile(
-                "CSE/D2CC/CRACs/crac-test",
-                MinioAdapterConstants.DEFAULT_GRIDCAPA_INPUT_GROUP_METADATA_VALUE,
-                "CRAC",
-                OffsetDateTime.parse("2021-10-01T21:00Z"),
-                OffsetDateTime.parse("2021-10-01T22:00Z"),
-                OffsetDateTime.now());
-        taskRepository.save(task);
-
-        Event eventCracDeletion = TaskManagerTestUtil.createEvent(minioAdapter, "CSE_D2CC", INPUT_FILE_GROUP_VALUE, "CRAC", "CSE/D2CC/CRACs/crac-test", "2021-09-30T21:00Z/2021-09-30T22:00Z");
-        taskManager.removeProcessFile(eventCracDeletion);
-
-        assertEquals(NOT_CREATED, taskRepository.findByTimestamp(taskTimestamp).orElseThrow().getStatus());
     }
 
     @Test
